backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
mysqlclient==2.2.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
